requirements.txt
APScheduler==3.11.0
asarPy==1.0.1
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
Deprecated==1.2.18
Flask==3.1.1
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.71.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
opentelemetry-api==1.33.1
opentelemetry-distro==0.54b1
opentelemetry-exporter-otlp==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-exporter-otlp-proto-http==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asyncio==0.54b1
opentelemetry-instrumentation-click==0.54b1
opentelemetry-instrumentation-dbapi==0.54b1
opentelemetry-instrumentation-flask==0.54b1
opentelemetry-instrumentation-grpc==0.54b1
opentelemetry-instrumentation-jinja2==0.54b1
opentelemetry-instrumentation-logging==0.54b1
opentelemetry-instrumentation-psycopg2==0.54b1
opentelemetry-instrumentation-requests==0.54b1
opentelemetry-instrumentation-sqlalchemy==0.54b1
opentelemetry-instrumentation-sqlite3==0.54b1
opentelemetry-instrumentation-threading==0.54b1
opentelemetry-instrumentation-urllib==0.54b1
opentelemetry-instrumentation-urllib3==0.54b1
opentelemetry-instrumentation-wsgi==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
packaging==25.0
protobuf==5.29.4
psycopg2-binary==2.9.10
requests==2.32.3
SQLAlchemy==2.0.41
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
